fix hint to discriminate the kind of win
It looks like the last part of the name of an acme win, is not the full
hostname, but only the first subdomain part of it. And since the hint to
detect whether a win is for a file or for a shell relies on the win
name, this change fixes that hint accordingly.